onnx/requirements.txt
flatbuffers==25.12.19
ml_dtypes==0.5.4
mpmath==1.3.0
numpy==2.4.2
onnx==1.20.0
onnxruntime-gpu==1.24.1
opencv-python==4.13.0.92
packaging==26.0
protobuf==6.33.5
sympy==1.14.0
typing_extensions==4.15.0
